sgp-compbench/requirements.txt
cairosvg==2.8.2
openai==1.102.0
Pillow==11.3.0
requests==2.32.5
tqdm==4.67.1
